+++ tests/test_actors.py
@@ -606,3 +606,8 @@ def test_workers_log_rate_limit_exceeded_errors_differently(stub_broker, stub_wo
         # Then warning mock should be called with a special message
         warning_messages = [args[0] for _, args, _ in warning_mock.mock_calls]
         assert "Rate limit exceeded in message %s: %s." in warning_messages
+
+
+def test_as_dict_actor(stub_broker, do_work):
+    res = do_work.as_dict()
+    assert res == {"name": "do_work", "priority": 0, "queue_name": "default"}